Move virtual_hashes codegen to own file
