+++ godot-codegen/src/generator/method_tables.rs
@@ -177,11 +177,6 @@ pub fn make_utility_function_table(api: &ExtensionApi) -> TokenStream {
     make_named_method_table(table)
 }
 
-#[cfg(since_api = "4.4")]
-pub fn make_virtual_hashes_table(api: &ExtensionApi) -> TokenStream {
-    make_virtual_hashes_for_all_classes(&api.classes)
-}
-
 // ----------------------------------------------------------------------------------------------------------------------------------------------
 // Implementation
 
@@ -303,52 +298,6 @@ fn make_named_method_table(info: NamedMethodTable) -> TokenStream {
                     #( #method_inits )*
                 }
             }
-        }
-    }
-}
-
-#[cfg(since_api = "4.4")]
-fn make_virtual_hashes_for_all_classes(all_classes: &[Class]) -> TokenStream {
-    let modules = all_classes
-        .iter()
-        .map(|class| make_virtual_hashes_for_class(class));
-
-    quote! {
-        #![allow(non_snake_case, non_upper_case_globals)]
-
-        #( #modules )*
-    }
-}
-
-#[cfg(since_api = "4.4")]
-fn make_virtual_hashes_for_class(class: &Class) -> TokenStream {
-    let class_rust_name = &class.name().rust_ty;
-
-    let constants: Vec<TokenStream> = class
-        .methods
-        .iter()
-        .filter_map(|method| {
-            let FnDirection::Virtual { hash } = method.direction() else {
-                return None;
-            };
-
-            let method_name = method.name_ident();
-            let constant = quote! {
-                pub const #method_name: u32 = #hash;
-            };
-
-            Some(constant)
-        })
-        .collect();
-
-    // Don't generate mod SomeClass {} without contents.
-    if constants.is_empty() {
-        return TokenStream::new();
-    }
-
-    quote! {
-        pub mod #class_rust_name {
-            #( #constants )*
         }
     }
 }
